Adding user role and user based Post's access

+++ api/ctrls/todo-ctrls.js
@@ -1,66 +1,93 @@
 const PostModel = require('../modals/TodoPost');
 
 const handlers = {};
-const restrictKeys = 'message addedDate complete success';
+const restrictKeys = 'author message addedDate complete success';
 
 handlers.getAll = (req, res, next) => {
-  PostModel.find({}, restrictKeys).then(
-    result => {
-      res.send({
-        list: result,
-        success: true
-      });
-    },
-    err => {
-      res.status(500).send(err);
-    }
-  );
-};
-
-handlers.getItem = (req, res, next) => {
-  let id = req.params.id;
-  PostModel.findById(id, restrictKeys).then(
-    result => {
-      if(result) {
+  const isAdmin = req.decoded.isAdmin;
+  if (isAdmin) {
+    PostModel.find({}, restrictKeys).then(
+      result => {
         res.send({
-          ...result.toJSON(),
+          posts: result,
           success: true
         });
-      } else {
-        res.status(404).send({
-          message: 'Item not found',
-          success: false
-        });
+      },
+      err => {
+        res.status(500).send(err);
       }
-    },
-    err => {
-      res.status(500).send({
-        message: 'Failed to find the item',
-        success: false
+    );
+  } else {
+    // return only user's post
+    next();
+  }
+};
+
+handlers.getUserPosts = (req, res, next) => {
+  const userId = req.decoded.id;
+  PostModel.find({ author: userId}, restrictKeys).then(
+    result => {
+      res.send({
+        posts: result,
+        success: true
       });
     }
-  )
+  ).catch(error => {
+    res.status(500).send({
+      message: error.message,
+      success: false
+    });
+  });
+}
+
+handlers.getItem = (req, res, next) => {
+  const userId = req.decoded.id;
+  const id = req.params.id;
+  PostModel.findOne({
+    '_id': id,
+    author: userId
+   }, restrictKeys).then(
+    result => {
+      res.send({
+        post: result,
+        success: true
+      });
+    }
+  ).catch(error => {
+    res.status(500).send({
+      message: 'Failed to find the item',
+      success: false
+    });
+  });
 }
 
 handlers.addItem = (req, res, next) => {
-  let newItem = new PostModel(req.body);
+  const userId = req.decoded.id;
+  const newItem = new PostModel({
+    ...req.body,
+    author: userId
+  });
   newItem.save().then(
     success => {
-      res.status(200).send({
-        result: success,
+      res.send({
+        post: success,
         message: 'Item added',
         success: true
       });
     },
-    err => {
-      res.status(500).send(err);
+    error => {
+      res.status(500).send({
+        message: error.message,
+        success: false
+      });
     }
   );
 }
 
 handlers.updateItem = (req, res, next) => {
-  let id = req.body.id;
-  PostModel.findOneAndUpdate({ '_id': id },
+  const userId = req.decoded ? req.decoded.id || '' : '';
+  const id = req.params.id;
+  PostModel.findOneAndUpdate({ '_id': id, author: userId },
     { $set: {
         message: req.body.message,
         complete: req.body.complete,
@@ -69,8 +96,8 @@ handlers.updateItem = (req, res, next) => {
   ).then (
     result => {
       if (result) {
-        res.status(200).send({
-          ...result.toJSON(),
+        res.send({
+          post: { ...result.toJSON() },
           success: true
         });
       } else {
@@ -91,11 +118,12 @@ handlers.updateItem = (req, res, next) => {
 }
 
 handlers.deleteItem = (req, res, next) => {
-  let id = req.body.id;
-  PostModel.deleteOne({ '_id': id }).then(
+  const userId = req.decoded ? req.decoded.id || '' : '';
+  const id = req.params.id;
+  PostModel.deleteOne({ '_id': id, author: userId }).then(
     result => {
       if(result.n) {
-        res.status(200).send({
+        res.send({
           message: 'Item deleted',
           success: true
         });
